Extract manual tests not run by default into their own file
No real changes, just move more tests in a separate file.
